# Conflicts:
#	app/src/main/java/ru/surf/learn2invest/network_components/responses/CoinPriceDto.kt

+++ app/src/main/java/ru/surf/learn2invest/network_components/responses/CoinReviewDto.kt
@@ -5,11 +5,11 @@ data class CoinReviewDto(
     val rank: Int, //ранг
     val symbol: String, //Абревиатура
     val name: String, //Название
-    val supply: Double,
-    val maxSupply: Double,
-    val marketCapUsd: Double, //Капитализация
-    val volumeUsd24Hr: Double,
-    val priceUsd: Double, //Цена
-    val changePercent24Hr: Double, //Изменение цены за последни 24 часа
-    val vwap24Hr: Double
+    val supply: Float,
+    val maxSupply: Float,
+    val marketCapUsd: Float, //Капитализация
+    val volumeUsd24Hr: Float,
+    val priceUsd: Float, //Цена
+    val changePercent24Hr: Float, //Изменение цены за последни 24 часа
+    val vwap24Hr: Float
 )
