+++ app/src/main/java/ru/surf/learn2invest/noui/database_components/entity/Transaction.kt
@@ -12,8 +12,8 @@ open class Transaction(
     @PrimaryKey(autoGenerate = true) var id: Int = 0,
     val name: String, //Имя коина (Bitcoin)
     val symbol: String, //Абревиатура (BTC)
-    val coinPrice: Double, //Закупочная цена одного
-    val dealPrice: Double, //Общая умма сделки (цена * количество)
-    val amount: Double, //Количество
+    val coinPrice: Float, //Закупочная цена одного
+    val dealPrice: Float, //Общая умма сделки (цена * количество)
+    val amount: Float, //Количество
     val transactionType: TransactionsType
 )
